pyproject.toml
[project]
name = "pybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "ipykernel>=7.1.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "pillow>=12.1.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "tzdata>=2025.3",
]
